requirements.txt
tensorflow-cpu
Werkzeug
Flask
numpy
Pillow
gunicorn
numpy
imageio